scraper/requirements.txt
playwright
beautifulsoup4
psycopg2-binary
pydantic
python-dotenv
